-"Recent Posts Widget" was created -"class-recent-posts-widget.php" was created & registered in "class-sidebars.php"
